added some tranches hooks
